completed the delete functionality

+++ components/Post.js
@@ -54,8 +54,10 @@ const Post = ({ post }) => {
   };
 
   const deletePost = async () => {
-    await deleteDoc(doc(db, "posts", post.id));
-    deleteObject(ref(storage, `posts/${post.id}/image`));
+    if (window.confirm("Are you sure you want to delete this post?")) {
+      await deleteDoc(doc(db, "posts", post.id));
+      deleteObject(ref(storage, `posts/${post.id}/image`));
+    }
   };
 
   return (
